Unify runtime protocol handling across Raspbot components
The PC, car, and app surfaces now share one protocol vocabulary while the coordinator records richer runtime status for startup, alarm, and bridge flows. The car-side protocol module is kept as a compatibility shim so existing imports continue to work while shared packet definitions move into raspbot_shared.

Constraint: User requested committing source-code updates only, excluding thesis, PPT, generated outputs, and demo/simulation tools.

Rejected: Commit generated thesis/PPT artifacts | user explicitly narrowed this commit to source code only.

Rejected: Include demo and simulation tools | user explicitly excluded those from this commit.

Confidence: medium

Scope-risk: moderate

Reversibility: clean

Tested: Reviewed staged name-status and diff stat before commit.

Not-tested: Full PC test suite and selected PC tests both exceeded the 120s command timeout in this environment.

+++ raspbot1/pc_modules/cry_detector.py
@@ -131,7 +131,7 @@ class YamnetCryDetector:
         if model is None or not self._cry_indices:
             return 0.0
         try:
-            scores, _embeddings, _spectrogram = model(waveform.astype(np.float32))
+            scores, _embeddings, _spectrogram = model(waveform)
             scores_np = scores.numpy() if hasattr(scores, "numpy") else np.asarray(scores)
             mean_scores = np.mean(scores_np, axis=0)
             return float(np.max(mean_scores[self._cry_indices]))
